fix(server): add context event properties
