+++ server/executor/trigger_result_processor_worker.go
@@ -128,10 +128,9 @@ func (r triggerResultProcessorWorker) handleExecutionResult(run test.Run, ctx co
 	if run.TriggerResult.Error != nil {
 		run = run.TriggerFailed(fmt.Errorf(run.TriggerResult.Error.ErrorMessage))
 
-		analytics.SendEvent("test_run_finished", "error", "", &map[string]string{
-			"finalState": string(run.State),
-			"tenant_id":  middleware.TenantIDFromContext(ctx),
-		})
+		eventData := map[string]string{"finalState": string(run.State)}
+		eventData = analytics.InjectProperties(eventData, middleware.EventPropertiesFromContext(ctx))
+		analytics.SendEvent("test_run_finished", "error", "", &eventData)
 
 		return run
 	}
